Create functions to help populate default values
Create populate_with_defaults! to complete user tables by appending
columns that have default values.

+++ src/model-preparation.jl
@@ -473,13 +473,18 @@ end
 
 function add_expressions_to_constraints!(connection, variables, constraints)
     # creating a workspace with enough entries for any of the representative periods or normal periods
-    maximum_num_timesteps = only(
-        row[1] for
-        row in DuckDB.query(connection, "SELECT MAX(num_timesteps) FROM rep_periods_data")
-    )::Int32
-    maximum_num_periods = only(
-        row[1] for row in DuckDB.query(connection, "SELECT MAX(period) FROM rep_periods_mapping")
-    )::Int32
+    maximum_num_timesteps = Int64(
+        only(
+            row[1] for
+            row in DuckDB.query(connection, "SELECT MAX(num_timesteps) FROM rep_periods_data")
+        ),
+    )
+    maximum_num_periods = Int64(
+        only(
+            row[1] for
+            row in DuckDB.query(connection, "SELECT MAX(period) FROM rep_periods_mapping")
+        ),
+    )
     Tmax = max(maximum_num_timesteps, maximum_num_periods)
     workspace = [Dict{Int,Float64}() for _ in 1:Tmax]
